+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # develop
+  * Feature: Allow registering temporary (single execution) handlers which are removed after they are triggered
 
 # 1.1.0 - 2012-01-21
   * Feature: Allow guarding on the value of method calls with arguments, by using an array as a hash key